+++ src/iceberg/CMakeLists.txt
@@ -21,10 +21,10 @@ set(ICEBERG_SOURCES
     arrow_c_data_internal.cc
     catalog/in_memory_catalog.cc
     demo.cc
-    datum.cc
     expression/expression.cc
     file_reader.cc
     json_internal.cc
+    literal.cc
     manifest_entry.cc
     manifest_list.cc
     metadata_columns.cc
